+++ common/upper.test.ts
@@ -1,0 +1,13 @@
+import { assertEquals } from "@std/assert";
+import { upper } from "./mod.ts";
+
+Deno.test("upper", () => {
+  assertEquals(upper.parse("Hello"), {
+    success: true,
+    results: [{
+      value: "H",
+      remaining: "ello",
+      position: { line: 1, column: 1 },
+    }],
+  });
+});
