+++ examples/common.ts
@@ -1,13 +1,3 @@
-import { alt } from "../combinators/alternation/alt.ts";
-import { between } from "../combinators/sequencing/between.ts";
-import { seq } from "../combinators/sequencing/seq.ts";
-import { parseErrors } from "../errors.ts";
-import { createParser, foldL1, type Parser, updatePosition } from "../index.ts";
-import { many } from "../combinators/iteration/many.ts";
-import { repeat } from "../combinators/iteration/repeat.ts";
-import { sepBy } from "../combinators/iteration/sepBy.ts";
-import { result } from "../primitives/mod.ts";
-import { regex } from "../common/regex.ts";
 
 /**
  * Represents a predicate function
@@ -37,234 +27,3 @@ type Predicate = (input: string) => boolean;
 export function regexPredicate(regex: RegExp): Predicate {
   return (input: string) => regex.test(input);
 }
-
-/**
- * Consumes the next character of the input and fails if the input is empty.
- *
- * @example Non-empty input
- *
- * ```ts
- * const { results } = take.parse("hello");
- * // [{value: 'h', remaining: 'ello', ...}]
- * ```
- *
- * @example Empty input
- *
- * ```ts
- * const { message } = take.parse("");
- * // "Unexpected end of input"
- * ```
- */
-export const take: Parser<string> = createParser(
-  (input, position) => {
-    if (input.length === 0) {
-      return {
-        success: false,
-        message: parseErrors.takeError,
-        position,
-      };
-    }
-
-    const consumed = input[0];
-    const newPosition = updatePosition(position, consumed);
-
-    return {
-      success: true,
-      results: [{
-        value: consumed,
-        position: newPosition,
-        remaining: input.slice(1),
-      }],
-    };
-  },
-);
-
-/**
- * Parses the next two characters
- */
-export const takeTwo: Parser<string> = repeat(take, 2).map((arr) =>
-  arr.join("")
-).error(parseErrors.takeTwoError);
-
-/**
- * Parses a single white space with the regex `/\s\/`
- *
- * @throws Throws "Expected a white space character" when the parse fails
- */
-export const whitespace: Parser<string> = regex(/^\s/).error(
-  parseErrors.whitespace,
-);
-
-/**
- * Parses whitespaces (0 or more)
- *
- * Regex: /\s*\/
- */
-export const whitespaces: Parser<string> = regex(/^\s*/);
-
-/**
- * Parses whitespaces (1 or more)
- *
- * Regex: /\s+\/
- */
-export const whitespaces1: Parser<string> = regex(/^\s+/);
-
-/**
- * Parses the space character (0 or more)
- */
-export const spaces: Parser<string> = regex(/^ */);
-
-/**
- * Parses the newline character
- */
-export const newline: Parser<string> = regex(/^\n/);
-
-/**
- * Matches against a specific character or keyword
- *
- * @example
- *
- * ```ts
- * const dot = literal(".");
- * const { results } = dot.parse(".23");
- * // [{value: '.', remaining: '23', ...}]
- *
- * const { message } = dot.parse("0.23");
- * // "Expected '.' but got '0'"
- * ```
- */
-export function literal(value: string): Parser<string> {
-  return (createParser((input, position) => {
-    if (!input.startsWith(value)) {
-      return {
-        success: false,
-        position,
-        message: `Expected '${value}', but got '${
-          input.slice(0, value.length) || "EOI"
-        }'`,
-      };
-    }
-
-    return {
-      success: true,
-      results: [{
-        value,
-        remaining: input.slice(value.length),
-        position: updatePosition(position, value),
-      }],
-    };
-  }));
-}
-
-/**
- * Parses a token and discards trailing spaces
- */
-export function token(value: string): Parser<string> {
-  return literal(value).skipTrailing(whitespaces);
-}
-
-/**
- * Parses a single letter (case insensitive)
- *
- * Regex: /^[a-zA-Z]/
- */
-export const letter: Parser<string> = regex(/^[a-zA-Z]/).error(
-  parseErrors.letter,
-);
-
-/**
- * Parses a string of letters
- */
-export const letters: Parser<string> = many(letter).map((letters) =>
-  letters.join("")
-);
-
-/**
- * Parses a single lower case letter
- *
- * * Regex: /^[a-z]/
- */
-export const lower: Parser<string> = regex(/^[a-z]/).error(parseErrors.lower);
-
-/**
- * Parses a single upper case letter
- *
- * * Regex: /^[A-Z]/
- */
-export const upper: Parser<string> = regex(/^[A-Z]/).error(parseErrors.upper);
-
-/**
- * Parses alphanumeric characters (0 or more)
- *
- * * Regex: /^\w*\/
- */
-export const alphaNums: Parser<string> = regex(/^\w*/);
-
-/**
- * Parses an identifier token as letter + alphanums
- */
-export const identifier: Parser<string> = letter.bind((l) =>
-  alphaNums.map((rest) => l + rest)
-);
-
-/**
- * Parses a single digit
- *
- * * Regex: /^\d/
- */
-export const digit: Parser<number> = regex(/^\d/).map(Number.parseInt).error(
-  parseErrors.digit,
-);
-
-/**
- * Parses a natural number
- */
-export const natural: Parser<number> = foldL1(
-  digit,
-  result((a: number, b: number) => 10 * a + b),
-).skipTrailing(spaces).error(parseErrors.natural);
-
-/**
- * Parses an integer (element of ℤ)
- */
-export const integer: Parser<number> = alt(
-  literal("-").bind(() => natural).map((x) => -x),
-  literal("+").bind(() => natural).map((x) => x),
-  natural,
-).error(parseErrors.integer);
-
-/**
- * Parses a decimal number aka a float
- */
-export const decimal: Parser<number> = seq(
-  integer,
-  literal("."),
-  natural,
-).map(([integral, _, fractional]) =>
-  integral +
-  Math.sign(integral) * Math.pow(10, -Math.ceil(Math.log10(fractional))) *
-    fractional
-).skipTrailing(spaces).error(parseErrors.decimal);
-
-/**
- * Parses a number as decimal | integer
- */
-export const number: Parser<number> = alt(decimal, integer).error(
-  parseErrors.number,
-);
-
-/**
- * Builds a parser that expects the list syntax [p(,p)*]
- */
-export function listOf<T>(parser: Parser<T>): Parser<T[]> {
-  return between(
-    token("["),
-    sepBy(parser, token(",")),
-    token("]"),
-  );
-}
-
-/**
- * Parses a list of integers
- */
-export const listOfInts: Parser<number[]> = listOf(integer);
