+++ primitives/mod.ts
@@ -5,4 +5,6 @@
  */
 
 export { result } from "./result.ts";
+export { take } from "./take.ts";
+export { takeTwo } from "./takeTwo.ts";
 export { zero } from "./zero.ts";
