+++ tests/common.test.ts
@@ -1,124 +1,14 @@
 import { assertEquals } from "@std/assert";
-import { parseErrors } from "../errors.ts";
-import {
-  decimal,
-  digit,
-  integer,
-  letter,
-  listOfInts,
-  literal,
-  lower,
-  natural,
-  regexPredicate,
-  take,
-  takeTwo,
-  upper,
-} from "../examples/common.ts";
+import { regexPredicate } from "../examples/common.ts";
+import { lower } from "../common/mod.ts";
+import { take } from "../primitives/mod.ts";
 import { filter } from "../index.ts";
-import { repeat } from "../combinators/iteration/repeat.ts";
+import { repeat } from "../combinators/iteration/mod.ts";
 
-Deno.test("item", () => {
-  assertEquals(take.parse(""), {
-    success: false,
-    message: parseErrors.takeError,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(take.parse("monad"), {
-    success: true,
-    results: [{
-      value: "m",
-      remaining: "onad",
-      position: { line: 1, column: 1 },
-    }],
-  });
-});
-
-Deno.test("two items", () => {
-  assertEquals(takeTwo.parse("m"), {
-    success: false,
-    message: parseErrors.takeTwoError,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(takeTwo.parse("monad"), {
-    success: true,
-    results: [{
-      value: "mo",
-      remaining: "nad",
-      position: { line: 1, column: 2 },
-    }],
-  });
-});
-
-Deno.test("letter", () => {
-  assertEquals(letter.parse("m"), {
-    success: true,
-    results: [{ value: "m", remaining: "", position: { line: 1, column: 1 } }],
-  });
-
-  assertEquals(letter.parse("1"), {
-    success: false,
-    message: parseErrors.letter,
-    position: { line: 1, column: 0 },
-  });
-});
-
-const twoLower = repeat(lower, 2).map((letters) => letters.join("")).error(
-  "Expected two lowercase letters",
-);
-
-Deno.test("lower", () => {
-  assertEquals(lower.parse("Hello"), {
-    success: false,
-    message: parseErrors.lower,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(twoLower.parse("abcd"), {
-    success: true,
-    results: [{
-      value: "ab",
-      remaining: "cd",
-      position: { line: 1, column: 2 },
-    }],
-  });
-
-  assertEquals(twoLower.parse("aBcd"), {
-    success: false,
-    message: "Expected two lowercase letters",
-    position: { line: 1, column: 0 },
-  });
-});
-
-Deno.test("upper", () => {
-  assertEquals(upper.parse("Hello"), {
-    success: true,
-    results: [{
-      value: "H",
-      remaining: "ello",
-      position: { line: 1, column: 1 },
-    }],
-  });
-});
-
-Deno.test("digit", () => {
-  assertEquals(digit.parse("a"), {
-    success: false,
-    message: parseErrors.digit,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(digit.parse("2"), {
-    success: true,
-    results: [{ value: 2, remaining: "", position: { line: 1, column: 1 } }],
-  });
-
-  assertEquals(digit.parse("23"), {
-    success: true,
-    results: [{ value: 2, remaining: "3", position: { line: 1, column: 1 } }],
-  });
-});
+export const twoLower = repeat(lower, 2).map((letters) => letters.join(""))
+  .error(
+    "Expected two lowercase letters",
+  );
 
 const even = filter(take, regexPredicate(/^[02468]/)).error(
   "Expected an even number",
@@ -138,144 +28,5 @@ Deno.test("filter even", () => {
     success: false,
     message: "Expected an even number",
     position: { line: 1, column: 0 },
-  });
-});
-
-Deno.test("literal", () => {
-  assertEquals(literal("m").parse("a"), {
-    success: false,
-    message: "Expected 'm', but got 'a'",
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(literal("m").parse("monad"), {
-    success: true,
-    results: [{
-      value: "m",
-      remaining: "onad",
-      position: { line: 1, column: 1 },
-    }],
-  });
-
-  assertEquals(literal("hello").parse("hello there"), {
-    success: true,
-    results: [{
-      value: "hello",
-      remaining: " there",
-      position: { line: 1, column: 5 },
-    }],
-  });
-
-  assertEquals(literal("hello").parse("helicopter"), {
-    success: false,
-    message: "Expected 'hello', but got 'helic'",
-    position: { line: 1, column: 0 },
-  });
-});
-
-Deno.test("natural", () => {
-  assertEquals(natural.parse("23 and more"), {
-    success: true,
-    results: [{
-      value: 23,
-      remaining: "and more",
-      position: { line: 1, column: 3 },
-    }],
-  });
-
-  assertEquals(natural.parse("and more"), {
-    success: false,
-    message: parseErrors.natural,
-    position: { line: 1, column: 0 },
-  });
-});
-
-Deno.test("integer", () => {
-  assertEquals(integer.parse("23 and more"), {
-    success: true,
-    results: [{
-      value: 23,
-      remaining: "and more",
-      position: { line: 1, column: 3 },
-    }],
-  });
-
-  assertEquals(integer.parse("and more"), {
-    success: false,
-    message: parseErrors.integer,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(integer.parse("-23 and more"), {
-    success: true,
-    results: [{
-      value: -23,
-      remaining: "and more",
-      position: { line: 1, column: 4 },
-    }],
-  });
-});
-
-Deno.test("decimal", () => {
-  assertEquals(decimal.parse("2.3 and more"), {
-    success: true,
-    results: [{
-      value: 2.3,
-      remaining: "and more",
-      position: { line: 1, column: 4 },
-    }],
-  });
-
-  assertEquals(decimal.parse("23 and more"), {
-    success: false,
-    message: parseErrors.decimal,
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(decimal.parse("-2.3 and more"), {
-    success: true,
-    results: [{
-      value: -2.3,
-      remaining: "and more",
-      position: { line: 1, column: 5 },
-    }],
-  });
-});
-
-Deno.test("list of integers", () => {
-  assertEquals(listOfInts.parse("[1 ,  -2, 3] and more"), {
-    success: true,
-    results: [{
-      value: [1, -2, 3],
-      remaining: "and more",
-      position: { line: 1, column: 13 },
-    }],
-  });
-
-  assertEquals(listOfInts.parse("[] and more"), {
-    success: true,
-    results: [{
-      value: [],
-      remaining: "and more",
-      position: { line: 1, column: 3 },
-    }],
-  });
-
-  assertEquals(listOfInts.parse("1 ,  -2, 3] and more"), {
-    success: false,
-    message: "Expected '[', but got '1'",
-    position: { line: 1, column: 0 },
-  });
-
-  assertEquals(listOfInts.parse("[a ,  -2, 3] and more"), {
-    success: false,
-    message: "Expected ']', but got 'a'",
-    position: { line: 1, column: 1 },
-  });
-
-  assertEquals(listOfInts.parse("[1  -2, 3] and more"), {
-    success: false,
-    message: "Expected ']', but got '-'",
-    position: { line: 1, column: 4 },
   });
 });
