+++ tests/index.test.ts
@@ -1,17 +1,15 @@
 import { assertEquals, assertIsError, assertThrows } from "@std/assert";
-import { any } from "../combinators/alternation/any.ts";
+import { any } from "../combinators/alternation/mod.ts";
 import { many } from "../combinators/iteration/many.ts";
 import { seq } from "../combinators/sequencing/seq.ts";
 import { ParseError, parseErrors } from "../errors.ts";
-import {
-  digit,
-  letter,
-  literal,
-  number,
-  take,
-  takeTwo,
-  whitespace,
-} from "../examples/common.ts";
+import { takeTwo } from "../primitives/takeTwo.ts";
+import { number } from "../common/number.ts";
+import { digit } from "../common/digit.ts";
+import { letter } from "../common/letter.ts";
+import { literal } from "../common/literal.ts";
+import { whitespace } from "../common/whitespace.ts";
+import { take } from "../primitives/take.ts";
 
 Deno.test("skipTrailing", () => {
   assertEquals(digit.skipTrailing(letter).parse("1a"), {
